+++ Bundle_DE.properties
@@ -1,0 +1,93 @@
+fileText=Datei
+editText=Bearbeiten
+formatText=Format
+viewText=Ansicht
+helpText=Hilfe
+changeText=Sprache
+windowNew=Neues Fenster
+fileNew=Neue Datei
+fileOpen=Datei \u00D6ffnen...
+fileSave=Datei speichern
+fileSaveAs=Datei speichern als...
+filePageSetup=Seiteneinstellungen...
+filePrint=Drucken
+fileExit=Beenden
+editUndo=R\u00FCckg\u00E4ngig
+editCut=Ausschneiden
+editCopy=Kopieren
+editPaste=Einf\u00FCgen
+editDelete=L\u00F6schen
+editFind=Suchen...
+editFindNext=N\u00E4chstes finden
+editReplace=Ersetzen
+editGoTo=Gehe zu...
+editSelectAll=Alles ausw\u00E4hlen
+editTimeDate=Zeit/Datum
+commandoopen=Terminal \u00F6ffnen
+formatWordWrap=Zeilenumbruch
+formatFont=Schrift...
+formatForeground=Textfarbe...
+formatBackground=Hintergrundfarbe...
+viewStatusBar=Statusleiste
+helpHelpTopic=Online Hilfe
+helpHelpoffline=Offline Hilfe
+helpAboutFNotepad=\u00DCber FNotepad
+aboutText=<html><big>FNotepad</big><hr><hr><p align=center>Von fantastic-octo-garbanzo!<hr><p align=center>Mit OpenJDK15 compiliert.<br><br><strong>Danke f\u00FCr das Benutzen von FNotepad!</strong><br>Bei Bugs und Ideen gerne ein Issue auf Github stellen<p align=center>
+Lang1=English
+Lang2=Italiano
+fileName=Unbenannt
+applicationName=FNotepad
+statusbar.init1=Tabulatorbreite: 
+statusbar.init2=\     ||      Zeichen: 0, W\u00F6rter: 0       ||       Zeile: 1, Spalte: 1  
+statusbar.work1=\     ||      Zeichen 
+statusbar.work2=, W\u00F6rter 
+statusbar.work3=\       ||       Zeile 
+statusbar.work4=, Spalte 
+goto=Zeile eingeben:
+search.text1=Nichts zu suchen, bitte Finden im Bearbeiten-Men\u00FC zuerst probieren !!!!
+font.text1=Schrift ausw\u00E4hlen
+about.text1=\u00DCber FNotepad
+implement.text1=Dieser Befehl wird gerade integriert
+help.URL=https://github.com/fantastic-octo-garbanzo/FNotepad/wiki/Hilfe
+help.File=/bin/Hilfe.html
+help.Temp=Hilfe
+FileOperation.open=\u00D6ffne Datei
+FileOperation.FileFilter=Alle Dateien
+LookandFeel=Aussehen \u00E4ndern
+FontChooser.font=Schrift
+FontChooser.typeface=Schriftart
+FontChooser.size=Gr\u00F6\u00DFe
+FontChooser.Ok=Best\u00E4tigen
+FontChooser.Cancel=Abbrechen
+FontChooser.fontStyles={"Normal","Kursiv","Fett","Fett Kursiv"}
+FindDialog.matchCase=Gro\u00DF/Kleinschreibung
+FindDialog.Up=Hoch
+FindDialog.Down=Runter
+FindDialog.Direction=Richtung
+FindDialog.Next=N\u00E4chstes suchen
+FindDialog.What=Was suchen 
+FileFilter.AllFiles=Alle Dateien(*.*)
+FindDialog.ReplaceW=Ersetzen mit 
+FindDialog.Total=Vollst\u00E4ndig ersetzt = 
+FindDialog.Error1=Konnte nicht gefunden werden\\"
+FindDialog.ErrorT=Suchen
+FindDialog.Search=Suchen
+FindDialog.Replace=Ersetzen
+FindDialog.ReplaceA=Alles ersetzen
+FileOperation.file=Datei
+FileOperation.save1=Speichern als...
+FileOperation.save2=Jetzt speichern
+FileOperation.save3=Hier speichern!
+FileOperation.save4=\ existiert schon.<br>Ersetzen?<html>
+FileOperation.save5=Speichern
+FileOperation.open1=\u00D6ffne Datei...
+FileOperation.open2=\u00D6ffnen
+FileOperation.open3=Ausgew\u00E4hlte Datei \u00F6ffnen.
+FileOperation.open4=<br>Datei nicht gefunden.<br>Bitte \u00FCberpr\u00FCfen Sie den angegebenen Dateinamen.<html>
+FileOperation.open5=\u00D6ffnen
+FileOperation.readonly=(Schreibgesch\u00FCtzt)
+FileOperation.readonly1=Erfolgreich gespeichert/ge\u00F6ffnet.
+FileOperation.readonly2=Fehler beim \u00D6ffnen/Speichern : 
+FileOperation.confirmSave1=<html>Der Inhalt der Datei \\"
+FileOperation.confirmSave2=\\" wurde ge\u00E4ndert.<br>Wollen Sie die \u00C4nderungen speichern?<html>
+
