+++ Bundle_EN.properties
@@ -1,0 +1,92 @@
+fileText=File
+editText=Edit
+formatText=Format
+viewText=View
+helpText=Help
+changeText=Language
+windowNew=New Window
+fileNew=New File
+fileOpen=Open File...
+fileSave=Save File
+fileSaveAs=Save File As...
+filePageSetup=Page Setup...
+filePrint=Print
+fileExit=Exit
+editUndo=Undo
+editCut=Cut
+editCopy=Copy
+editPaste=Paste
+editDelete=Delete
+editFind=Find...
+editFindNext=Find Next
+editReplace=Replace
+editGoTo=Go To...
+editSelectAll=Select All
+editTimeDate=Time/Date
+commandoopen=Open Terminal
+formatWordWrap=Word Wrap
+formatFont=Font...
+formatForeground=Set Text color...
+formatBackground=Set Pad color...
+viewStatusBar=Status Bar
+helpHelpTopic=Online Help
+helpHelpoffline=Offline Help
+helpAboutFNotepad=About FNotepad
+aboutText=<html><big>FNotepad</big><hr><hr><p align=center>By fantastic-octo-garbanzo!<hr><p align=left>Compiled with OpenJDK 15.<br><br><strong>Thank you for using FNotepad!</strong><br>For ideas and bug reports<br>feel free to create a issue on Github!<p align=center>
+Lang1=Deutsch
+Lang2=Italiano
+fileName=Unnamed
+applicationName=FNotepad
+statusbar.init1=Tabulatorsize: 
+statusbar.init2=\     ||      Letters 0, Words 0       ||       Ln 1, Col 1  
+statusbar.work1=\       ||      Letters 
+statusbar.work2=, Words 
+statusbar.work3=\       ||       Line 
+statusbar.work4=, Column 
+goto=Enter Line Number:
+search.text1=Nothing to search for, use Find option of Edit Menu first !!!!
+font.text1=Choose a font
+about.text1=About FNotepad
+implement.text1=This command is yet to be implemented
+help.URL=https://github.com/fantastic-octo-garbanzo/FNotepad/wiki/Help
+help.File=/bin/Help.html
+help.Temp=Help
+FileOperation.open=Open File
+FileOperation.FileFilter=All Files
+LookandFeel=Change Look and Feel
+FontChooser.font=Font
+FontChooser.typeface=Font Style
+FontChooser.size=Size
+FontChooser.Ok=OK
+FontChooser.Cancel=Cancel
+FontChooser.fontStyles="Regular","Italic","Bold","Bold Italic"
+FindDialog.matchCase=Match case
+FindDialog.Up=Up
+FindDialog.Down=Down
+FindDialog.Direction=Direction
+FindDialog.Next=Find Next
+FindDialog.What=Find what 
+FileFilter.AllFiles=All Files(*.*)
+FindDialog.ReplaceW=Replace With 
+FindDialog.Total=Total replacements made = 
+FindDialog.Error1=Cannot find\\"
+FindDialog.ErrorT=Find
+FindDialog.Search=Search
+FindDialog.Replace=Replace
+FindDialog.ReplaceA=Replace All
+FileOperation.file=File
+FileOperation.save1=Save as...
+FileOperation.save2=Save Now
+FileOperation.save3=Click me to save!
+FileOperation.save4=\ already exists.<br>Do you want to replace it?<html>
+FileOperation.save5=Save As
+FileOperation.open1=Open File...
+FileOperation.open2=Open this
+FileOperation.open3=Click me to open the selected file.!
+FileOperation.open4=<br>file not found.<br>Please verify the correct file name was given.<html>
+FileOperation.open5=Open
+FileOperation.readonly=(Read only)
+FileOperation.readonly1=Saved/opened successfully.
+FileOperation.readonly2=Failed to save/open : 
+FileOperation.confirmSave1=<html>The text in the 
+FileOperation.confirmSave2=\ file has been changed.<br>Do you want to save the changes?<html>
