+++ Bundle_IT.properties
@@ -1,0 +1,92 @@
+fileText=file
+editText=Modificare
+formatText=formato
+viewText=Visualizza
+helpText=Aiuto
+changeText=lingua
+windowNew=Nuova finestra
+fileNew=nuovo file
+fileOpen=Apri il file...
+fileSave=salvare il file
+fileSaveAs=Salva file con nome...
+filePageSetup=Impostazioni della pagina...
+filePrint=Pressare
+fileExit=rottura
+editUndo=Annullato
+editCut=ritagliare
+editCopy=copia
+editPaste=Inserire
+editDelete=Chiaro
+editFind=Ricerca...
+editFindNext=Trova il prossimo
+editReplace=Sostituire
+editGoTo=Vai a...
+editSelectAll=Seleziona tutto
+editTimeDate=Ora dell'appuntamento
+commandoopen=Apri terminale
+formatWordWrap=nuova linea
+formatFont=font...
+formatForeground=Colore del testo...
+formatBackground=Colore di sfondo
+viewStatusBar=Barra di stato
+helpHelpTopic=Aiuto online
+helpHelpoffline=assistenza offline
+helpAboutFNotepad=Tramite FNotepad
+aboutText=<html><big>FNotepad</big><hr><hr><p align=center>Da fantastico-octo-garbanzo!<hr><p align=center>Compilato con OpenJDK15.<br><br><strong>Grazie per aver utilizzato FNotepad!</strong><br>Se hai bug e idee, pubblica un problema su Github<p align=center>
+Lang1=English
+Lang2=Deutsch
+fileName=Senza nome
+applicationName=FNotepad
+statusbar.init1=Larghezza tab: 
+statusbar.init2=\     ||      personaggio: 0, parole: 0       ||       riga: 1, colonna: 1  
+statusbar.work1=\     ||      personaggio 
+statusbar.work2=, parole 
+statusbar.work3=\       ||       personaggio 
+statusbar.work4=, colonna 
+goto=Immettere la riga:
+search.text1=Niente da cercare, prova prima a trovarlo nel menu di modifica !!!!
+font.text1=Seleziona il carattere
+about.text1=Tramite FNotepad
+implement.text1=Questo comando è attualmente in fase di integrazione
+help.URL=https://github.com/fantastic-octo-garbanzo/FNotepad/wiki/Aiuto
+help.File=/bin/Aiuto.html
+help.Temp=Aiuto
+FileOperation.open=Aprire il file
+FileOperation.FileFilter=All Files
+LookandFeel=Cambiare aspetto
+FontChooser.font=Scrittura
+FontChooser.typeface=Font
+FontChooser.size=Dimensione
+FontChooser.Ok=Confermare
+FontChooser.Cancel=Cancella
+FontChooser.fontStyles="Normale","Corsivo","Grasso","Grassetto italico"
+FindDialog.matchCase=Maiuscolo/minuscolo
+FindDialog.Up=Alto
+FindDialog.Down=Scendi
+FindDialog.Direction=Direzione
+FindDialog.Next=Ricerca successiva
+FindDialog.What=Cosa cercare 
+FileFilter.AllFiles=Tutti i file(*.*)
+FindDialog.ReplaceW=Sostituire con 
+FindDialog.Total=Total replacements made = 
+FindDialog.Error1=Non può essere trovato\\"
+FindDialog.ErrorT=Cerca
+FindDialog.Search=Cerca
+FindDialog.Replace=Sostituire
+FindDialog.ReplaceA=Sostituire tutti
+FileOperation.file=file
+FileOperation.save1=Salva con nome...
+FileOperation.save2=Salva ora
+FileOperation.save3=Clicca su di me per salvare!
+FileOperation.save4=\ esiste già.<br>Vuoi sostituirlo?<html>
+FileOperation.save5=Salva con nome
+FileOperation.open1=Aprire il file...
+FileOperation.open2=Aprire questo
+FileOperation.open3=Clicca su me per aprire il file selezionato.!
+FileOperation.open4=<br>file non trovato.<br>Si prega di verificare che il nome del file sia corretto.<html>
+FileOperation.open5=Aprire
+FileOperation.readonly=(Leggere solo)
+FileOperation.readonly1=Salvato/aperto con successo.
+FileOperation.readonly2=Impossibile salvare/aprire : 
+FileOperation.confirmSave1=<html>Il testo nel 
+FileOperation.confirmSave2=\ è stato cambiato.<br>Vuoi salvare le modifiche?<html>
