+++ FileOperation/FileOperationLANG.java
@@ -4,59 +4,59 @@ import java.io.*;
 import java.awt.event.*;
 import javax.swing.*;
 import FileFilter.*;
-import NotePads.FNotepadDE;
+import NotePads.FNotepadLANG;
 
-class FileOperationExampleDE extends JFrame {
+class FileOperationExampleLANG extends JFrame {
     JMenuBar mb;
-	JMenu file;
-	JMenuItem open;
-	JTextArea ta;
-	FileOperationExampleDE(){
-		open = new JMenuItem("\u00D6ffne Datei");
-		file = new JMenu("Datei");
-		file.add(open);
-		mb = new JMenuBar();
-		mb.setBounds(0,0,800,20);
-		mb.add(file);
-		ta = new JTextArea(800,800);
-		ta.setBounds(0,20,800,800);
-		add(mb);
-		add(ta);
-	}      
-	
-	public void actionPerformed(ActionEvent e) {
-		if(e.getSource() == open) {
-			JFileChooser fc = new JFileChooser();
-			int i = fc.showOpenDialog(this);
-			if(i == JFileChooser.APPROVE_OPTION) {
-				File f = fc.getSelectedFile();
-				String filepath = f.getPath();
-				try {
-					BufferedReader br = new BufferedReader(new FileReader(filepath));
-					String s1 = "",s2 = "";
-					while((s1 = br.readLine()) != null){
-						s2+=s1+"\n";
-					}
-					ta.setText(s2);
-					br.close();
-				}catch (Exception ex) {ex.printStackTrace();}
-			}
-		}
+    JMenu file;
+    JMenuItem open;
+    JTextArea ta;
+    FileOperationExampleLANG(){
+        open = new JMenuItem("\u00D6ffne Datei");
+        file = new JMenu("Datei");
+        file.add(open);
+        mb = new JMenuBar();
+        mb.setBounds(0,0,800,20);
+        mb.add(file);
+        ta = new JTextArea(800,800);
+        ta.setBounds(0,20,800,800);
+        add(mb);
+        add(ta);
     }
-///////////////////////////////////////
+
+    public void actionPerformed(ActionEvent e) {
+        if(e.getSource() == open) {
+            JFileChooser fc = new JFileChooser();
+            int i = fc.showOpenDialog(this);
+            if(i == JFileChooser.APPROVE_OPTION) {
+                File f = fc.getSelectedFile();
+                String filepath = f.getPath();
+                try {
+                    BufferedReader br = new BufferedReader(new FileReader(filepath));
+                    String s1 = "",s2 = "";
+                    while((s1 = br.readLine()) != null){
+                        s2+=s1+"\n";
+                    }
+                    ta.setText(s2);
+                    br.close();
+                }catch (Exception ex) {ex.printStackTrace();}
+            }
+        }
+    }
+    ///////////////////////////////////////
     public static void main(String[] args) {
-		new FileOperationExampleDE();
-	}
+        new FileOperationExampleLANG();
+    }
 }
 /*************************************/
 // Beginn der Klasse FileOperationDE
-public class FileOperationDE {
-    FNotepadDE npd;
+public class FileOperationLANG {
+    FNotepadLANG npd;
 
     public static boolean saved;
     boolean newFileFlag;
     static String fileName;
-    String applicationTitle = "FNotepad";
+    String applicationTitle = this.npd.bundle.getString("applicationName");
 
     File fileRef;
     JFileChooser chooser;
@@ -81,34 +81,34 @@ public class FileOperationDE {
         this.fileName = new String(fileName);
     }
     /////////////////////////
-    public FileOperationDE(FNotepadDE npd) {
+    public FileOperationLANG(FNotepadLANG npd) {
         this.npd = npd;
 
         saved = true;
         newFileFlag = true;
-        fileName = new String("Unbenannt");
+        fileName = this.npd.bundle.getString("fileName");
         fileRef = new File(fileName);
         this.npd.f.setTitle(fileName + " - " + applicationTitle);
 
-		// Verschiedene Dateiendungen
+        // Verschiedene Dateiendungen
         chooser = new JFileChooser();
-        chooser.addChoosableFileFilter(new FileFilterDE("*", "Alle Dateien"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".txt", "Text Files(*.txt)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".java", "Java Source Files(*.java)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".py", "Python Files(*.py)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".c", "C Programming Language(*.c)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".cpp", "C++(*.cpp)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".cs", "C#(*.cs)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".d", "D Programming Language(*.d)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".sh", "Shell Script File(*.sh)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".bat", "Batch File(*.bat)"));
-		chooser.addChoosableFileFilter(new FileFilterDE(".rtf", "Rich Text Format(*.rtf)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".pdf", "Portable Document Files(*.pdf)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".html", "Hyper Text Markup Language(*.html)"));
-        chooser.addChoosableFileFilter(new FileFilterDE(".asm", "Assembler(*.asm)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG("*", this.npd.bundle.getString("FileOperation.FileFilter")));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".txt", "Text Files(*.txt)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".java", "Java Source Files(*.java)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".py", "Python Files(*.py)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".c", "C Programming Language(*.c)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".cpp", "C++(*.cpp)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".cs", "C#(*.cs)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".d", "D Programming Language(*.d)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".sh", "Shell Script File(*.sh)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".bat", "Batch File(*.bat)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".rtf", "Rich Text Format(*.rtf)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".pdf", "Portable Document Files(*.pdf)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".html", "Hyper Text Markup Language(*.html)"));
+        chooser.addChoosableFileFilter(new FileFilterLANG(".asm", "Assembler(*.asm)"));
         chooser.setCurrentDirectory(new File("."));
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     boolean saveFile(File temp) {
         try {
             fout = new FileWriter(temp);
@@ -125,33 +125,33 @@ public class FileOperationDE {
         updateStatus(temp, true);
         return true;
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     public boolean saveThisFile() {
         if (!newFileFlag) {
             return saveFile(fileRef);
         }
         return saveAsFile();
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     public boolean saveAsFile() {
-        chooser.setDialogTitle("Speichern als...");
-        chooser.setApproveButtonText("Jetzt speichern");
+        chooser.setDialogTitle(this.npd.bundle.getString("FileOperation.save1"));
+        chooser.setApproveButtonText(this.npd.bundle.getString("FileOperation.save2"));
         chooser.setApproveButtonMnemonic(KeyEvent.VK_S);
-        chooser.setApproveButtonToolTipText("Hier speichern!");
+        chooser.setApproveButtonToolTipText(this.npd.bundle.getString("FileOperation.save3"));
         do {
             if (chooser.showSaveDialog(this.npd.f) != JFileChooser.APPROVE_OPTION)
                 return false;
             temp = chooser.getSelectedFile();
             if (!temp.exists()) break;
             if (JOptionPane.showConfirmDialog(
-                    this.npd.f, "<html>" + temp.getPath() + " existiert schon.<br>Ersetzen?<html>",
-                    "Speichern", JOptionPane.YES_NO_OPTION
+                    this.npd.f, "<html>" + temp.getPath() + this.npd.bundle.getString("FileOperation.save4"),
+                    this.npd.bundle.getString("FileOperation.save5"), JOptionPane.YES_NO_OPTION
             ) == JOptionPane.YES_OPTION)
                 break;
         } while (true);
         return saveFile(temp);
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     boolean openFile(File temp) {
         try {
             fin = new FileInputStream(temp);
@@ -178,13 +178,13 @@ public class FileOperationDE {
         this.npd.ta.setCaretPosition(0);
         return true;
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     public void openFile() {
         if (!confirmSave()) return;
-        chooser.setDialogTitle("\u00D6ffne Datei...");
-        chooser.setApproveButtonText("\u00D6ffnen");
+        chooser.setDialogTitle(this.npd.bundle.getString("FileOperation.open1"));
+        chooser.setApproveButtonText(this.npd.bundle.getString("FileOperation.open2"));
         chooser.setApproveButtonMnemonic(KeyEvent.VK_O);
-        chooser.setApproveButtonToolTipText("Ausgew\u00E4hlte Datei \u00F6ffnen.");
+        chooser.setApproveButtonToolTipText(this.npd.bundle.getString("FileOperation.open3"));
 
         do {
             if (chooser.showOpenDialog(this.npd.f) != JFileChooser.APPROVE_OPTION)
@@ -194,16 +194,15 @@ public class FileOperationDE {
             if (temp.exists()) break;
 
             JOptionPane.showMessageDialog(this.npd.f,
-                    "<html>" + temp.getName() + "<br>Datei nicht gefunden.<br>" +
-                            "Bitte \u00FCberpr\u00FCfen Sie den angegebenen Dateinamen.<html>",
-                    "Öffnen", JOptionPane.INFORMATION_MESSAGE);
+                    "<html>" + temp.getName() + this.npd.bundle.getString("FileOperation.open4"),
+                    this.npd.bundle.getString("FileOperation.open5"), JOptionPane.INFORMATION_MESSAGE);
 
         } while (true);
 
         this.npd.ta.setText("");
 
         if (!openFile(temp)) {
-            fileName = "Unbenannt";
+            fileName = this.npd.bundle.getString("fileName");
             saved = true;
             this.npd.f.setTitle(fileName + " - " + applicationTitle);
         }
@@ -211,27 +210,26 @@ public class FileOperationDE {
             newFileFlag = true;
 
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     void updateStatus(File temp, boolean saved) {
         if (saved) {
             this.saved = true;
             fileName = new String(temp.getName());
             if (!temp.canWrite()) {
-                fileName += "(Schreibgesch\u00FCtzt)";
+                fileName += this.npd.bundle.getString("FileOperation.readonly");
                 newFileFlag = true;
             }
             fileRef = temp;
             npd.f.setTitle(fileName + " - " + applicationTitle);
-            npd.statusBar.setText("Datei : " + temp.getPath() + " erfolgreich gespeichert/ge\u00F6ffnet.");
+            npd.statusBar.setText(this.npd.bundle.getString("FileOperation.readonly1"));
             newFileFlag = false;
         } else {
-            npd.statusBar.setText("Fehler beim \u00D6ffnen/Speichern : " + temp.getPath());
+            npd.statusBar.setText(this.npd.bundle.getString("FileOperation.readonly2") + temp.getPath());
         }
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     public boolean confirmSave() {
-        String strMsg = "<html>Der Inhalt der Datei \"" + fileName + "\" wurde ge\u00E4ndert.<br>" +
-                "Wollen Sie die \u00C4nderungen speichern?<html>";
+        String strMsg = this.npd.bundle.getString("FileOperation.confirmSave1") + fileName + this.npd.bundle.getString("FileOperation.confirmSave2");
         if (!saved) {
             int x = JOptionPane.showConfirmDialog(this.npd.f, strMsg, applicationTitle, JOptionPane.YES_NO_CANCEL_OPTION);
 
@@ -240,11 +238,11 @@ public class FileOperationDE {
         }
         return true;
     }
-//////////////////////////////////////
+    //////////////////////////////////////
     public void newFile() {
         if (!confirmSave()) return;
         this.npd.ta.setText("");
-        fileName = new String("Unbenannt");
+        fileName = this.npd.bundle.getString("fileName");
         fileRef = new File(fileName);
         saved = true;
         newFileFlag = true;
