--- MenuConstants/MenuConstantsDE.java
@@ -1,54 +1,0 @@
-package MenuConstants;
-
-public interface MenuConstantsDE {
-    String fileText = "Datei";
-    String editText = "Bearbeiten";
-    String formatText = "Format";
-    String viewText = "Ansicht";
-    String helpText = "Hilfe";
-    String changeText = "Sprache";
-
-    String windowNew = "Neues Fenster";
-    String fileNew = "Neue Datei";
-    String fileOpen = "Datei \u00D6ffnen...";
-    String fileSave = "Datei speichern";
-    String fileSaveAs = "Datei speichern als...";
-    String filePageSetup = "Seiteneinstellungen...";
-    String filePrint = "Drucken";
-    String fileExit = "Beenden";
-
-    String editUndo = "R\u00FCckg\u00E4ngig";
-    String editCut = "Ausschneiden";
-    String editCopy = "Kopieren";
-    String editPaste = "Einf\u00FCgen";
-    String editDelete = "L\u00F6schen";
-    String editFind = "Suchen...";
-    String editFindNext = "N\u00E4chstes finden";
-    String editReplace = "Ersetzen";
-    String editGoTo = "Gehe zu...";
-    String editSelectAll = "Alles ausw\u00E4hlen";
-    String editTimeDate = "Zeit/Datum";
-    String commandoopen = "Terminal \u00F6ffnen";
-
-    String formatWordWrap = "Zeilenumbruch";
-    String formatFont = "Schrift...";
-    String formatForeground = "Textfarbe...";
-    String formatBackground = "Hintergrundfarbe...";
-
-    String viewStatusBar = "Statusleiste";
-
-    String helpHelpTopic = "Online Hilfe";
-    String helpHelpoffline = "Offline Hilfe";
-    String helpHelpOnline = "Github-Projektseite";
-    String helpAboutFNotepad = "\u00DCber FNotepad";
-
-    String aboutText =
-            "<html><big>FNotepad</big><hr><hr>"
-                    + "<p align=center>Von fantastic-octo-garbanzo!"
-                    + "<hr><p align=center>Mit OpenJDK15 compiliert.<br><br>"
-                    + "<strong>Danke f\u00FCr das Benutzen von FNotepad!</strong><br>"
-                    + "Bei Bugs und Ideen gerne ein Issue auf Github stellen<p align=center>";
-
-    String LangEN = "English";
-    String LangIT = "Italiano";
-}
