--- MenuConstants/MenuConstantsEN.java
@@ -1,53 +1,0 @@
-package MenuConstants;
-
-public interface MenuConstantsEN {
-    String fileText = "File";
-    String editText = "Edit";
-    String formatText = "Format";
-    String viewText = "View";
-    String helpText = "Help";
-    String changeText = "Language";
-
-    String windowNew = "New Window";
-    String fileNew = "New File";
-    String fileOpen = "Open File...";
-    String fileSave = "Save File";
-    String fileSaveAs = "Save File As...";
-    String filePageSetup = "Page Setup...";
-    String filePrint = "Print";
-    String fileExit = "Exit";
-
-    String editUndo = "Undo";
-    String editCut = "Cut";
-    String editCopy = "Copy";
-    String editPaste = "Paste";
-    String editDelete = "Delete";
-    String editFind = "Find...";
-    String editFindNext = "Find Next";
-    String editReplace = "Replace";
-    String editGoTo = "Go To...";
-    String editSelectAll = "Select All";
-    String editTimeDate = "Time/Date";
-    String commandoopen = "Open Terminal";
-
-    String formatWordWrap = "Word Wrap";
-    String formatFont = "Font...";
-    String formatForeground = "Set Text color...";
-    String formatBackground = "Set Pad color...";
-
-    String viewStatusBar = "Status Bar";
-
-    String helpHelpTopic = "Online Help";
-    String helpHelpoffline = "Offline Help";
-    String helpAboutFNotepad = "About FNotepad";
-
-    String aboutText =
-            "<html><big>FNotepad</big><hr><hr>"
-                    + "<p align=center>By fantastic-octo-garbanzo!"
-                    + "<hr><p align=left>Compiled with OpenJDK 15.<br><br>"
-                    + "<strong>Thank you for using FNotepad!</strong><br>"
-                    + "For ideas and bug reports<br>"
-                    + "feel free to create a issue on Github!<p align=center>";
-    String LangDE = "Deutsch";
-    String LangIT = "Italiano";
-}
