--- MenuConstants/MenuConstantsIT.java
@@ -1,53 +1,0 @@
-package MenuConstants;
-
-public interface MenuConstantsIT {
-    String fileText = "file";
-    String editText = "Modificare";
-    String formatText = "formato";
-    String viewText = "Visualizza";
-    String helpText = "Aiuto";
-    String changeText = "lingua";
-
-    String windowNew = "Nuova finestra";
-    String fileNew = "nuovo file";
-    String fileOpen = "Apri il file...";
-    String fileSave = "salvare il file";
-    String fileSaveAs = "Salva file con nome...";
-    String filePageSetup = "Impostazioni della pagina...";
-    String filePrint = "Pressare";
-    String fileExit = "rottura";
-
-    String editUndo = "Annullato";
-    String editCut = "ritagliare";
-    String editCopy = "copia";
-    String editPaste = "Inserire";
-    String editDelete = "Chiaro";
-    String editFind = "Ricerca...";
-    String editFindNext = "Trova il prossimo";
-    String editReplace = "Sostituire";
-    String editGoTo = "Vai a...";
-    String editSelectAll = "Seleziona tutto";
-    String editTimeDate = "Ora dell'appuntamento";
-    String commandoopen = "Apri terminale";
-
-    String formatWordWrap = "nuova linea";
-    String formatFont = "font...";
-    String formatForeground = "Colore del testo...";
-    String formatBackground = "Colore di sfondo...";
-
-    String viewStatusBar = "Barra di stato";
-
-    String helpHelpTopic = "Aiuto online";
-    String helpHelpoffline = "assistenza offline";
-    String helpAboutFNotepad = "Tramite FNotepad";
-
-    String aboutText =
-            "<html><big>FNotepad</big><hr><hr>"
-                    + "<p align=center>Da fantastico-octo-garbanzo!"
-                    + "<hr><p align=center>Compilato con OpenJDK15.<br><br>"
-                    + "<strong>Grazie per aver utilizzato FNotepad!</strong><br>"
-                    + "Se hai bug e idee, pubblica un problema su Github<p align=center>";
-
-    String LangEN = "English";
-    String LangDE = "Deutsch";
-}
